+++ nix/mypy-nox.py
@@ -33,6 +33,9 @@ def gen_stubs(session: nox.Session) -> None:
     session.install("-r", "requirements.txt", "-r", "dev-requirements.txt")
     session.run("stubgen", *FILES, "-o", ".", "--include-private", "--no-import")
 
+    _paths = [p + 'i' for p in FILES]
+    session.run("isort", *_paths)
+    session.run("black", *_paths)
 
 @nox.session(reuse_venv=True)
 def type_check(session: nox.Session) -> None: